Added time formatting to the script
Localtime returned by api was in year first format, which was really
ugly, so I formatted it to be dd/mm/yyyy, and I added a fxn that
checks if it's the morning or evening. It's in 24h time, I'll roll
with it for now, but I might end up changing it later.