Add dom manipulation logic to new code
Yeah so I didn't like the way my new code was set up, so
I made yet ANOTHER file. This one is set up for webpack, and
the modules actually do the thing. Added the src and dist folders,
with config files. The forecast divs know when you click on them and
they'll soon be able to populate the hourly display with their data.