Add initial styling, date logic for forecast
When you load the forecast, now it will give you the day of the week
that the forecast is for, instead of the ugly localtime in
yyyy-mm-dd that it originally returns. Marks the current day of the
week as 'Today', and then keeps going from there. Also, each of the
'current' displays have their own section now instead of just overwriting
the text content of the div selected.